+++ backend/src/main/java/com/planmanager/www/model/users/dto/UserResponseDTO.java
@@ -2,6 +2,13 @@ package com.planmanager.www.model.users.dto;
 
 import com.planmanager.www.model.users.UserRole;
 
-public record UserResponseDTO(String id, String completeName, String email, UserRole role, Long prefeituraId) {
+public record UserResponseDTO
+        (
+                String id,
+                String completeName,
+                String email,
+                UserRole role,
+                Integer prefeituraId
+        ) {
 
 }
